have the TileLoader return QImages rather than TextureTiles
This patch fixes a bug introduced in 8101c98e, where a wrong blending may be assigned if a texture layer is used in multiple map themes, causing the map to stay black.
This happens if sourceDir() from different GeoSceneTextures are the same, which then map to the same hash value in the TileId.
As a result, only one GeoSceneTexture "survives" in the TileLoader, such that the blending from the wrong GeoSceneTexture may be applied to a TextureTile.
Having the TileLoader return a plain QImage and having the TextureTile created in StackedTileLoader solves the issue, since in the correct GeoSceneTexture is used in the StackedTileLoader.

As a nice side effect, MergedLayerDecorator::loadDataset() doesn't need to rely on QSharedPointer to delete the temporarily used TextureTile.