Put occupants of HOVERS_ABOVE_SURFACE layer into a defined z-order
Both the GeometryLayer and the PlacemarkLayout are rendered in the layer
HOVERS_ABOVE_SURFACE, but neither had a zValue set, so defaulting to 0.0

According to LayerInterface.h:62: "If both have the same z value, their
paint order is undefined." By chance, the PlacemarkLayout was always
drawn above the GeometryLayer and everything worked out just fine.

When the Elevation Profile's "red flag thingy" came along as a third
player on the HOVERS_ABOVE_SURFACE layer, the dice fell the other way
around and suddenly the placemarks were rendered beneath the geometry.

Thanks for the help, Thibaut!
BUG: 288662
FIXED-IN: 1.3.0